Evaluation-Agent/eval_agent/eval_tools/t2i_comp/requirements.txt
git+https://github.com/openai/CLIP.git
# git+https://github.com/facebookresearch/detectron2.git@5aeb252b194b93dc2879b4ac34bc51a31b5aee13
absl-py==1.4.0
accelerate==0.17.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==23.1.0
black==22.3.0
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.4
confection==0.0.4
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
datasets==2.13.0
dill==0.3.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
exceptiongroup==1.1.1
fairscale==0.4.4
filelock==3.12.2
fire==0.5.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.20.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
huggingface-hub==0.15.1
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
iopath==0.1.9
Jinja2==3.1.2
kiwisolver==1.4.4
langcodes==3.3.0
lit==16.0.6
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
pathy==0.10.2
Pillow==9.5.0
platformdirs==3.8.0
pluggy==1.2.0
portalocker==2.7.0
preshed==3.0.8
protobuf==4.23.3
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pydantic==1.10.9
pydot==1.4.2
Pygments==2.15.1
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
thinc==8.1.10
timm==0.4.12
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.2
wcwidth==0.2.6
Werkzeug==2.3.6
xxhash==3.2.0
yacs==0.1.8
yarl==1.9.2
zipp==3.15.0
